typing refactoring
