+++ ui-ngx/src/app/modules/home/components/widget/lib/gateway/connectors-configuration/configuration-json/models/configuration-validate.models.ts
@@ -1,0 +1,10 @@
+import { Ace } from 'ace-builds';
+
+export interface CustomAnnotation extends Ace.Annotation {
+  custom?: boolean;
+}
+
+export interface ConnectorConfigValidation {
+  valid: boolean;
+  annotations: CustomAnnotation[];
+}
